starter/integrate-services/event-grid/python/client/requirements.txt
azure-identity==1.21.0
azure-eventgrid==4.21.0
flask==3.1.1
